Working in dev-2 except highcharts display differs between export stats and dashboard
