+++ assets/build.sh
@@ -1,2 +1,2 @@
-npx esbuild "assets/scripts/**/common.js" --entry-names=[dir]/[name] --outbase=assets/scripts --bundle --minify --outdir=public/scripts $*
+npx esbuild "assets/scripts/**/main.js" --entry-names=[dir]/[name] --outbase=assets/scripts --bundle --minify --outdir=public/scripts $*
 #npx esbuild "assets/scripts/**/review.js" --entry-names=[dir]/[name] --outbase=assets/scripts --bundle --minify --outdir=public/scripts $*
