+++ cgi/adminUserRate.tt
@@ -50,9 +50,9 @@
 <br/>
 <input type="submit" value="Submit"/>
 <input type="button" value="Clear"
-       onclick="selMenuItem('projectSel', 0);
-                selMenuItem('yearSel', 0);
-                selMenuItem('institutionSel', 0);"
+       onclick="crms.selectOption('projectSel', 0);
+                crms.selectOption('yearSel', 0);
+                crms.selectOption('institutionSel', 0);"
 />
 
 </form>
@@ -63,7 +63,7 @@
   <div id="[% divid %]" style="[% style %]">
     <span style="font-size:1.3em;"><strong>[% data.title %]</strong></span>
     &nbsp;&nbsp;&nbsp;&nbsp;
-    <button type="button" onclick='toggleVisibility("RTDIV_[% param.id %]");'>
+    <button type="button" onclick='crms.toggleVisibility("RTDIV_[% param.id %]");'>
       Show Raw Data
     </button>
     <br/>
@@ -94,11 +94,11 @@
     [% data = crms.CreateUserStatsReport(param.user, param.year, param.proj, !param.old) %]
     [% IF param.old %]
       <a class="nowrap" style="display:block;margin-bottom:-18px;padding:4px;width:190px;color:#FFFFFF;background-color:#000000;cursor:pointer;text-align:center;"
-         onclick='toggleVisibility("[% divid %]");'>Show [% crms.GetProjectRef(param.proj).name _ ' ' _ param.year %]</a><br/>
+         onclick='crms.toggleVisibility("[% divid %]");'>Show [% crms.GetProjectRef(param.proj).name _ ' ' _ param.year %]</a><br/>
     [% END %]
     <div id="[% divid %]" style="[% style %]">
       <h3>[% data.title %]&nbsp;&nbsp;&nbsp;&nbsp;
-          <button type="button" onclick='toggleVisibility("RTDIV_[% param.id %]");'>
+          <button type="button" onclick='crms.toggleVisibility("RTDIV_[% param.id %]");'>
             Show Raw Data
           </button>
       </h3>
