+++ cgi/dashboard.tt
@@ -3,11 +3,9 @@
   <head>
     <meta charset="utf-8"/>
     <link rel="icon" type="image/png" href="[% crms.WebPath('web', crms.GetSystemVar('favicon')) %]"/>
-    <link rel="stylesheet" type="text/css" href="[% crms.WebPath('web', 'css/dashboard.css') %]"/>
+    <link rel="stylesheet" type="text/css" href="[% crms.WebPath('public', 'styles/dashboard.css') %]"/>
     <title>CRMS Dashboard</title>
-    <script type="text/javascript" src="[% crms.WebPath('web', 'js/common.js') %]"></script>
-    <script type="text/javascript" src="https://code.highcharts.com/adapters/standalone-framework.js"></script>
-    <script type="text/javascript" src="https://code.highcharts.com/highcharts.js"></script>
+    <script type="text/javascript" src="[% crms.WebPath('public', 'scripts/main.js') %]"></script>
 </head>
 <body>
 
@@ -17,16 +15,12 @@
 	{
 	  try
 	  {
-      loadChart('chart0', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=4') %]");
-      loadChart('chart1', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=0') %]");
-      loadChart('chart2', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=7') %]");
-      loadChart('chart3', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=2') %]");
-[% IF crms.get('sys') == 'crmsworld' %]
-      loadChart('chart4', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=5') %]");
-[% ELSE %]
-      loadChart('chart4', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=6') %]");
-[% END %]
-      loadChart('chart5', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=9') %]");
+      crms.loadChart('chart0', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=4') %]");
+      crms.loadChart('chart1', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=0') %]");
+      crms.loadChart('chart2', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=7') %]");
+      crms.loadChart('chart3', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=2') %]");
+      crms.loadChart('chart4', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=6') %]");
+      crms.loadChart('chart5', "[% crms.WebPath('cgi', 'getExportStats?type=graph&c=9') %]");
     }
     catch (e)
     {
