+++ cgi/home.tt
@@ -1,9 +1,5 @@
 [% INCLUDE header.tt %]
 
-<script type="text/javascript" src="https://code.highcharts.com/adapters/standalone-framework.js"></script>
-<script type="text/javascript" src="https://code.highcharts.com/highcharts.js"></script>
-<script type="text/javascript" src="https://code.highcharts.com/highcharts-more.js"></script>
-<script type="text/javascript" src="https://code.highcharts.com/modules/solid-gauge.js"></script>
 <script type="text/javascript">
 <!--
 function CheckAll(box,formid)
