+++ cgi/partial/authorities.tt
@@ -4,7 +4,7 @@
     <tr>
       <th class="nowrap">View 1:</th>
       <td>
-        <select title="Search 1" class="review" id="search1Select" onchange="changeFrame1(1);" style="float:right;">
+        <select title="Search 1" class="review" id="search1Select" onchange="crms.changeFrame1(1);" style="float:right;">
         [% FOREACH authority IN authorities %]
           <option value="[% authority.url %]"
                   [% IF authority.initial == 1 %]selected="selected"[% END %]>
@@ -17,7 +17,7 @@
     <tr>
       <th class="nowrap">View 2:</th>
       <td>
-        <select title="Search 2" class="review" id="search2Select" onchange="changeFrame2(1);" style="float:right;">
+        <select title="Search 2" class="review" id="search2Select" onchange="crms.changeFrame2(1);" style="float:right;">
         [% FOREACH authority IN authorities %]
           <option value="[% authority.url %]"
                   [% IF authority.initial == 2 %]selected="selected"[% END %]>
@@ -30,7 +30,7 @@
     <tr>
       <td></td>
       <td style="margin:0px;padding:0px;">
-         <input type="button" class="reviewblue" value="Toggle View" onclick="flipFrame();" accesskey="o"
+         <input type="button" class="reviewblue" value="Toggle View" onclick="crms.flipFrame();" accesskey="o"
          style="margin:0px;padding:1px;background-color:#e0d6de;float:right;"/>
       </td>
     </tr>
@@ -40,18 +40,17 @@
 [% FOREACH authority IN authorities %]
   [% IF authority.accesskey %]
     <a style="position:absolute;left:-999px;" href="#" accesskey="[% authority.accesskey %]"
-       onfocus="document.getElementById('search1Select').selectedIndex=[% i %];changeFrame1(1);">.</a>
+       onfocus="document.getElementById('search1Select').selectedIndex=[% i %];crms.changeFrame1(1);">.</a>
   [% END %]
   [% i = i + 1 %]
 [% END %]
 
   <script type="text/javascript">
 <!--
-  addEvent(window, 'load', Partial_authorities_mainWindowLoad);
-  function Partial_authorities_mainWindowLoad(e)
-  {
-    changeFrame1();
-    changeFrame2();
+  crms.addEvent(window, 'load', Partial_authorities_mainWindowLoad);
+  function Partial_authorities_mainWindowLoad(e) {
+    crms.changeFrame1();
+    crms.changeFrame2();
   }
 -->
   </script>
