+++ cgi/partial/copyrightForm.tt
@@ -140,7 +140,7 @@
         <input type="checkbox" id="hold" name="hold"
                [% (holds >=5 && !hold)? 'disabled="disabled"':'' %]
                [% (hold)? 'checked="checked"':'' %]
-               onclick="toggleVisibility('expiry');"/>
+               onclick="crms.toggleVisibility('expiry');"/>
       </td>
       <td><label for="hold">Hold for Question</label></td>
     </tr>
@@ -186,7 +186,7 @@
 <script>
 var gReviewData = [% data.json %];
 [% IF importUser %]
-  addEvent(window, 'load', function(e)
+  crms.addEvent(window, 'load', function(e)
   {
     popReviewInfo('[% importUser %]');
   });
@@ -210,7 +210,7 @@ function popReviewInfo(user)
   if (button) { button.checked = (renNum != null); }
   document.submitReview.renDate.value = renDate;
   document.submitReview.note.value = review.note;
-  selMenuItem('catMenu', (review.category)? review.category:'');
+  crms.selectOption('catMenu', (review.category)? review.category:'');
 }
 </script>
