+++ cgi/selection.tt
@@ -117,13 +117,13 @@ records per page.
 <br/>
 <input type="submit" value="Submit"/>
 <input type="button" value="Clear"
-       onclick="selMenuItem('op1','AND');
-                selMenuItem('op2','AND');
-                selMenuItem('search1','Identifier');
-                selMenuItem('search2','Identifier');
-                selMenuItem('search3','Identifier');
-                selMenuItem('order','[% defaultOrder %]');
-                selMenuItem('dir','[% defaultDir %]');
+       onclick="crms.selectOption('op1','AND');
+                crms.selectOption('op2','AND');
+                crms.selectOption('search1','Identifier');
+                crms.selectOption('search2','Identifier');
+                crms.selectOption('search3','Identifier');
+                crms.selectOption('order','[% defaultOrder %]');
+                crms.selectOption('dir','[% defaultDir %]');
                 document.getElementById('start-date').value='';
                 document.getElementById('end-date').value='';
                 document.getElementById('search1value').value='';
